+++ src/main/java/insane96mcp/enhancedai/mixin/MeleeAttackGoalMixin.java
@@ -1,29 +1,38 @@
 package insane96mcp.enhancedai.mixin;
 
 import insane96mcp.enhancedai.modules.base.feature.Attacking;
+import insane96mcp.insanelib.base.Feature;
+import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
-import net.minecraftforge.common.ForgeMod;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(MeleeAttackGoal.class)
-public class MeleeAttackGoalMixin {
+public abstract class MeleeAttackGoalMixin {
 
 	@Final
 	@Shadow
 	protected PathfinderMob mob;
 
-	@Inject(at = @At(value = "RETURN"), method = "getAttackReachSqr", cancellable = true)
-	public void getAttackReachSqr(LivingEntity livingEntity, CallbackInfoReturnable<Double> callbackInfo) {
-		if (!Attacking.shouldChangeAttackRange())
+	@Shadow protected abstract boolean isTimeToAttack();
+
+	@Shadow protected abstract void resetAttackCooldown();
+
+	@Inject(at = @At(value = "HEAD"), method = "checkAndPerformAttack", cancellable = true)
+	public void getAttackReachSqr(LivingEntity attacked, double distanceSqr, CallbackInfo ci) {
+		if (!Feature.isEnabled(Attacking.class))
 			return;
-		double attackRange = this.mob.getAttributeValue(ForgeMod.ATTACK_RANGE.get());
-		callbackInfo.setReturnValue(attackRange * attackRange + livingEntity.getBbWidth());
+		if (Attacking.isWithinMeleeAttackRange(this.mob, attacked) && this.isTimeToAttack() && this.mob.getSensing().hasLineOfSight(attacked)) {
+			this.resetAttackCooldown();
+			this.mob.swing(InteractionHand.MAIN_HAND);
+			this.mob.doHurtTarget(attacked);
+		}
+		ci.cancel();
 	}
 }